Don't make the arraysize() macro cast to int.

+++ re2/testing/exhaustive_tester.cc
@@ -174,7 +174,7 @@ void EgrepTest(int maxatoms, int maxops, const std::string& alphabet,
                const std::string& wrapper) {
   const char* tops[] = { "", "^(?:%s)", "(?:%s)$", "^(?:%s)$" };
 
-  for (int i = 0; i < arraysize(tops); i++) {
+  for (size_t i = 0; i < arraysize(tops); i++) {
     ExhaustiveTest(maxatoms, maxops,
                    Split("", alphabet),
                    RegexpGenerator::EgrepOps(),
