+++ util/util.h
@@ -5,7 +5,7 @@
 #ifndef UTIL_UTIL_H_
 #define UTIL_UTIL_H_
 
-#define arraysize(array) (int)(sizeof(array)/sizeof((array)[0]))
+#define arraysize(array) (sizeof(array)/sizeof((array)[0]))
 
 #ifndef ATTRIBUTE_NORETURN
 #if defined(__GNUC__)
